# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "ATWT API for Testing",
        "description": "This is an example api for testing purpose. automatedtestingwithtuyen.com",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "contact": {
            "name": "Tuyen Nguyen",
            "url": "http://automatedtestingwithtuyen.com",
            "email": "trongtuyen96@gmail.com"
        }
    },
    "host": "atwt-api.herokuapp.com",
    "basePath": "/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/user/{userId}/cars": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Find all cars of user by Id",
                "description": "Returns cars belongs to user",
                "operationId": "getAllCarsOfUserById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of user to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Car"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/car": {
            "get": {
                "tags": [
                    "car"
                ],
                "summary": "Get cars with conditions",
                "description": "",
                "operationId": "getCars",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "type",
                        "description": "Type of car",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "model",
                        "description": "Model of car",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "color",
                        "description": "Color of car",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "speed",
                        "description": "Speed of car (greater than or equal)",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Car"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "car"
                ],
                "summary": "Add a new car",
                "description": "",
                "operationId": "postCar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Car object that need to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/car/{carId}": {
            "get": {
                "tags": [
                    "car"
                ],
                "summary": "Find car by Id",
                "description": "Returns a single car",
                "operationId": "getCarById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "description": "Id of car to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Car not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "car"
                ],
                "summary": "Update an existing car",
                "description": "",
                "operationId": "putCar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "description": "Id of car to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Car object that needs to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Update successfully"
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Car not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "car"
                ],
                "summary": "Delete a car",
                "description": "",
                "operationId": "deleteCar",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "description": "Car Id to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Car not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Car": {
            "type": "object",
            "required": [
                "userID"
            ],
            "properties": {
                "type": {
                    "type": "string"
                },
                "speed": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Car"
            }
        },
...
    }
}